API coding part 3

+++ src/platform_power.c
@@ -35,10 +35,11 @@ void platform_power_poweroff()
 uint8_t platform_power_get_led()
 {
 	// status: 0 off, 1 on, 2 breathing, 3 flashing ?
-	uint8_t status = 0;
+	// Stack smashing it isn't a uint32_t, what?
+	uint32_t status = 0;
 	fpp_power_get_led_status(2, &status);
 	printf("LED Status: %d\n", status);
-	return status;
+	return (uint8_t)status;
 }
 
 // LED 0 (standby led) doesn't exist on my hardware...
